everything works as intended but ESLint has warning about useSearchParams hook
