+++ src/components/Pagination.tsx
@@ -6,17 +6,10 @@ import { range } from "../utils/helper";
 
 interface PaginationProps {
   count: number;
-  onPrevious: () => void;
-  onNext: () => void;
   fetchPeople: (endpoint?: string, term?: string, query?: string) => void;
 }
 
-const Pagination = ({
-  count,
-  onPrevious,
-  onNext,
-  fetchPeople,
-}: PaginationProps) => {
+const Pagination = ({ count, fetchPeople }: PaginationProps) => {
   const [activePage, setActivePage] = useState(1);
   const { pageQuery, updatePage } = useContext(GlobalContext);
   const amountOfPages = Math.ceil(count / 10);
@@ -32,10 +25,28 @@ const Pagination = ({
     fetchPeople("", "", `?page=${pageNumber}`);
   };
 
+  const previousClickHandler = () => {
+    if (activePage > 1) {
+      // fetchPeople(data.previous);
+      fetchPeople("", "", `?page=${pageQuery - 1}`);
+      updatePage(pageQuery - 1);
+      setActivePage((prev) => prev - 1);
+    }
+  };
+
+  const nextClickHandler = () => {
+    if (activePage < amountOfPages) {
+      // fetchPeople(data.previous);
+      fetchPeople("", "", `?page=${pageQuery + 1}`);
+      updatePage(pageQuery + 1);
+      setActivePage((prev) => prev + 1);
+    }
+  };
+
   return (
     <div className={classes["pagination-wrapper"]}>
       <div className={classes.pagination}>
-        <Button name="Previous" onClick={onPrevious} />
+        <Button name="Previous" onClick={previousClickHandler} />
         {slicedPagesRange.map((pageNum) => (
           <button
             className={classes["pages-btn"]}
@@ -50,7 +61,7 @@ const Pagination = ({
             {pageNum}
           </button>
         ))}
-        <Button name="Next" onClick={onNext} />
+        <Button name="Next" onClick={nextClickHandler} />
       </div>
     </div>
   );
